Add files via upload
Version 1 of the program in windows forms C#, it basically only has code, very basic. You can add, remove a service, that can only be done if you have an contact created; you can see contacts that you have, you can also add, remove and edit contacts from your list. This program saves your data in 2 files; 1 file is for the service part and the other is for the contacts.